pyproject.toml
[project]
name = "spotify-wrapped"
version = "0.1.0"
description = "Spotify Listening Dashboard"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "pandas>=2.2.3",
    "progress>=1.6",
    "streamlit>=1.53.1",
    "altair>=5.4.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.0",
    "mypy>=1.0.0",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
exclude = [
    "scripts/",
]
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "streamlit.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "altair.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dashboard", "scripts"]
include = [
    "app.py",
    "main.py",
]
